requirements/requirements.txt
selenium
python-dotenv
flask
requests
webdriver_manager